Tmux: set screen-256color as default, optionally upgrading to screen-256color-s
I.e. apparently Tmux doesn't evaluate if-shell commands early enough, as
a result the first window gets the default set outside of an if-shell
command. To ensure the colour scheme is at least close enough we set the
least functional terminal type outside of if-shell as default. Then when
we detect support for a more functional terminal type we upgrade the
default, that'll only apply to the next window but is at least better
than nothing.